Faz a modularização da função em arquivo separado
