+++ main.rb
@@ -1,3 +1,5 @@
+require_relative '/estrategias/funcao_remove_duplicatas'
+
 def executar_estrategia(lista, estrategia)
   puts estrategia.call(lista)
 end
@@ -16,14 +18,6 @@ def ordenacao_decrescente(lista)
         end
     end
   end
-end
-
-def remove_duplicatas lista
-  mapa = Hash.new 0
-  lista.each do |i|
-    mapa[i] += 1
-  end
-  print mapa.keys.inspect
 end
 
 lista = [11,5,3,1,2,3,12,3,32,2,1,12,3,2,1,56,64,21]
